backend/model_loader.py
"""
Model wrapper for loading a TensorFlow/Keras model and predicting on images.
Falls back to a placeholder predictor if TensorFlow or the model file is not available.
"""
from __future__ import annotations

import os
import hashlib
from typing import List, Tuple

import numpy as np
from PIL import Image

# Try to import TensorFlow/Keras. If unavailable, we will operate in fallback mode.
try:
    import tensorflow as tf  # type: ignore
    from tensorflow.keras.models import load_model  # type: ignore
except Exception:  # pragma: no cover - handled gracefully
    tf = None  # type: ignore
    load_model = None  # type: ignore


class ModelWrapper:
    """Simple wrapper around a Keras model with image preprocessing and label info."""

    def __init__(self, model_path: str, labels: List[str] | None = None, input_size: Tuple[int, int] = (224, 224)):
        self.model_path = model_path
        self.labels = labels or [
            'Healthy',
            'Powdery Mildew',
            'Leaf Spot',
        ]
        self.input_size = input_size  # (width, height)
        self.model = None
        self.loaded = False
        self.load_error = None

        # Attempt to load the model if possible
        if load_model is not None and os.path.exists(self.model_path):
            try:
                self.model = load_model(self.model_path, compile=False)
                self.loaded = True
                # Try to infer input size from the model if available
                try:
                    ishape = getattr(self.model, 'input_shape', None)
                    if ishape and len(ishape) == 4:
                        # Expecting (None, H, W, C)
                        h = ishape[1] if ishape[1] is not None else self.input_size[1]
                        w = ishape[2] if ishape[2] is not None else self.input_size[0]
                        if isinstance(w, int) and isinstance(h, int):
                            self.input_size = (int(w), int(h))
                except Exception:
                    pass
            except Exception as e:  # pragma: no cover - environment-specific
                self.model = None
                self.loaded = False
                self.load_error = str(e)
        else:
            self.loaded = False
            self.load_error = 'TensorFlow not available or model file missing.'

    # ---------------------------- Public API ---------------------------------
    def predict_image(self, image_path: str) -> Tuple[str, float]:
        """Predict the class label and confidence (0..1) for a given image path."""
        probs = self.predict_proba(image_path)
        idx = int(np.argmax(probs))
        conf = float(np.max(probs))
        label = self._label_from_index(idx)
        return label, conf

    def predict_proba(self, image_path: str) -> np.ndarray:
        """Return probability distribution over labels for the image (shape: [N])."""
        x = self._preprocess_image(image_path)

        if self.loaded and self.model is not None:
            preds = self.model.predict(x, verbose=0)
            if preds.ndim == 2:
                probs = preds[0]
            elif preds.ndim == 1:
                probs = preds
            else:
                probs = np.squeeze(preds)
            probs = probs.astype('float64')
            # Normalize to probabilities if needed
            s = float(np.sum(probs))
            if not (0.99 <= s <= 1.01):
                e = np.exp(probs - np.max(probs))
                probs = e / np.sum(e)
            return probs

        # Fallback: generate deterministic probabilities using Dirichlet
        idx, _ = self._fallback_predict(image_path)
        # Seed again to generate a full distribution deterministically
        seed = self._seed_from_file(image_path)
        rng = np.random.RandomState(seed)
        probs = rng.dirichlet(np.ones(len(self.labels)) * 2.0)
        # Ensure the chosen idx is top-1 to feel realistic
        probs[idx] = max(probs[idx], 0.5)
        probs = probs / probs.sum()
        return probs

    def get_info_for_label(self, label: str) -> str:
        """Return placeholder info text for a given label."""
        info_map = {
            'Healthy': 'The plant appears healthy. Maintain regular watering and nutrient schedule. Monitor for any early signs of stress.',
            'Powdery Mildew': 'Fungal disease causing white powdery spots. Improve air circulation, avoid overhead watering, and consider fungicidal treatment.',
            'Leaf Spot': 'Characterized by small dark lesions on leaves. Remove affected leaves and apply appropriate fungicide if spread continues.'
        }
        return info_map.get(label, 'No additional information available for this label.')

    def analyze_image_quality(self, image_path: str) -> dict:
        """Basic image quality analysis: blur (variance of Laplacian) and exposure."""
        with Image.open(image_path) as img:
            gray = img.convert('L')
            g = np.asarray(gray, dtype=np.float32)  # 0..255
        # Simple 2D Laplacian with edge padding
        lap = (-4*g)
        lap += np.pad(g, ((0,0),(1,0)), mode='edge')[:, :-1]
        lap += np.pad(g, ((0,0),(0,1)), mode='edge')[:, 1:]
        lap += np.pad(g, ((1,0),(0,0)), mode='edge')[:-1, :]
        lap += np.pad(g, ((0,1),(0,0)), mode='edge')[1:, :]
        var_lap = float(np.var(lap))
        blur_verdict = 'sharp' if var_lap >= 50.0 else 'blurry'

        brightness = float(np.mean(g))
        if brightness < 60:
            exposure_verdict = 'underexposed'
        elif brightness > 190:
            exposure_verdict = 'overexposed'
        else:
            exposure_verdict = 'well‑exposed'

        tips = []
        if blur_verdict == 'blurry':
            tips.append('Hold the camera steady and move closer until the leaf veins are crisp.')
        if exposure_verdict != 'well‑exposed':
            tips.append('Shoot in good natural light and avoid strong backlight.')

        verdict = 'Good' if blur_verdict == 'sharp' and exposure_verdict == 'well‑exposed' else 'Needs better focus/lighting'
        return {
            'blur_score': var_lap,
            'blur_verdict': blur_verdict,
            'brightness': brightness,
            'exposure_verdict': exposure_verdict,
            'verdict': verdict,
            'tips': ' '.join(tips)
        }

    # ---------------------------- Internals ----------------------------------
    def _label_from_index(self, idx: int) -> str:
        if 0 <= idx < len(self.labels):
            return self.labels[idx]
        return 'Unknown'

    def _preprocess_image(self, image_path: str) -> np.ndarray:
        """Load image, convert to RGB, resize to model input, scale to 0..1, add batch dim."""
        with Image.open(image_path) as img:
            img = img.convert('RGB')
            img = img.resize(self.input_size, Image.BILINEAR)
            arr = np.asarray(img).astype('float32') / 255.0
            arr = np.expand_dims(arr, axis=0)  # (1, H, W, C)
            return arr

    def _fallback_predict(self, image_path: str) -> Tuple[int, float]:
        """Deterministic placeholder based on image content hash to pick a label and confidence."""
        seed = self._seed_from_file(image_path)
        rng = np.random.RandomState(seed)
        idx = int(rng.randint(0, len(self.labels)))
        conf = float(rng.uniform(0.72, 0.97))
        return idx, conf

    def _seed_from_file(self, image_path: str) -> int:
        try:
            with open(image_path, 'rb') as f:
                digest = hashlib.sha1(f.read(1024 * 1024)).hexdigest()
        except Exception:
            digest = hashlib.sha1(image_path.encode('utf-8')).hexdigest()
        return int(digest[:8], 16) & 0xFFFFFFFF
