backend/requirements.txt
# Web
Flask>=2.2
Werkzeug>=2.2

# ML stack
numpy>=1.24
Pillow>=10.0

# Install TensorFlow for real inference; comment this line for placeholder-only mode
tensorflow>=2.12
